requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.23.2
python-dotenv==1.1.0
requests==2.32.0